- Working Reflection, moving towards SOLID - Refactored and moved reflection related traits and others inside the traits of the helpers
