+++ app/shared/io/helpers/ExtractHelper.scala
@@ -1,0 +1,5 @@
+package shared.io.helpers
+
+import shared.io.helpers.traits.ExtractHelperBase
+
+object ExtractHelper extends ExtractHelperBase
