+++ app/shared/io/helpers/ParallelTaskHelper.scala
@@ -1,0 +1,5 @@
+package shared.io.helpers
+
+import shared.io.helpers.traits.ParallelTaskHelperBase
+
+object ParallelTaskHelper extends ParallelTaskHelperBase
