+++ app/shared/io/helpers/ParameterCompareHelper.scala
@@ -1,0 +1,5 @@
+package shared.io.helpers
+
+import shared.io.helpers.traits.ParameterCompareHelperBase
+
+object ParameterCompareHelper extends ParameterCompareHelperBase
