+++ app/shared/io/helpers/ReflectionHelper.scala
@@ -1,8 +1,9 @@
 package shared.io.helpers
 
+import shared.io.helpers.implementation.NiceObject
 import shared.io.helpers.traits._
+import shared.io.helpers.traits.reflection.{ ClassTagHelper, ReflectionHelperBase }
 
-class ClassTagHelperConcreteImplementation extends ClassTagHelper
 
 object ReflectionHelper extends ReflectionHelperBase {
   implicit def toNiceObject[T <: AnyRef](x : T) : NiceObject[T] = new NiceObject(x)
