+++ app/shared/io/helpers/implementation/NiceObject.scala
@@ -1,4 +1,4 @@
-package shared.io.helpers.traits
+package shared.io.helpers.implementation
 
 class NiceObject[T <: AnyRef](x : T) {
   def niceClass : Class[_ <: T] = x.getClass.asInstanceOf[Class[T]]
