+++ app/shared/io/helpers/implementation/reflection/ClassTagHelperConcreteImplementation.scala
@@ -1,0 +1,5 @@
+package shared.io.helpers.implementation.reflection
+
+import shared.io.helpers.traits.reflection.ClassTagHelper
+
+class ClassTagHelperConcreteImplementation extends ClassTagHelper
