+++ app/shared/io/helpers/traits/MapHelperBase.scala
@@ -1,5 +1,7 @@
 package shared.io.helpers.traits
 
+import shared.io.helpers.implementation.HashMapWithArrayBufferAdderConcreteImplementation
+
 trait MapHelperBase {
   lazy val hashMapWithArrayBufferAdder : HashMapWithArrayBufferAdder = new HashMapWithArrayBufferAdderConcreteImplementation
 }
