+++ app/shared/io/helpers/traits/reflection/ClassTagHelperMemberInfoGetter.scala
@@ -1,7 +1,8 @@
-package shared.io.helpers.traits
+package shared.io.helpers.traits.reflection
 
 import shared.com.ortb.model.reflection.{ ClassMembersInfoBaseImplementationModel, ClassMembersInfoModel }
-import shared.io.helpers.ClassTagHelperConcreteImplementation
+import shared.io.helpers.implementation.reflection.ClassTagHelperConcreteImplementation
+import shared.io.helpers.traits.ExtractHelperBase
 
 import scala.reflect.ClassTag
 
@@ -12,9 +13,9 @@ trait ClassTagHelperMemberInfoGetter {
     val fields = getFieldWrapperModelsAsMap(classT)
     val methods = getMethodWrapperModelsAsMap(classT)
     val constructors = getConstructorWrapperModelsAsMap[T](ct)
-    val extractedFields = ExtractHelper.getFromResult(fields)
-    val extractedMethods = ExtractHelper.getFromResult(methods)
-    val extractedConstructors = ExtractHelper.getFromResult(constructors)
+    val extractedFields = ExtractHelperBase.getFromResult(fields)
+    val extractedMethods = ExtractHelperBase.getFromResult(methods)
+    val extractedConstructors = ExtractHelperBase.getFromResult(constructors)
 
     val clx = ClassMembersInfoBaseImplementationModel(
       classT,
